made search work better
